feat(error): add `ScaffdogAggregateError` custom error class
