+++ .changeset/short-students-smash.md
@@ -1,0 +1,12 @@
+---
+'@scaffdog/error': patch
+'scaffdog': patch
+---
+
+Add `ScaffdogAggregateError` custom error class.
+
+```typescript
+import { ScaffdogAggregateError } from '@scaffdog/error';
+// or
+// import { ScaffdogAggregateError } from 'scaffdog';
+```
